+++ package.json
@@ -35,6 +35,7 @@
   "dependencies": {
     "@phosphor-icons/react": "2.1.7",
     "@radix-ui/react-dialog": "^1.1.14",
+    "@radix-ui/react-dropdown-menu": "^2.1.15",
     "@radix-ui/react-label": "^2.1.7",
     "@radix-ui/react-select": "^2.2.5",
     "@recogito/standoff-converter": "^0.1.10",
